cscv2025/final-b/web/Emoji/requirements.txt
flask
duckdb
ftfy
llama-cpp-python
gunicorn
